SYNC-1569 Move to internal package
