+++ modules/apps/sync/sync-service/src/main/resources/com/liferay/sync/internal/upgrade/v1_0_2/dependencies/update.sql
