+++ modules/apps/sync/sync-service/src/main/resources/com/liferay/sync/internal/upgrade/v1_0_3/dependencies/update.sql
