+++ modules/apps/sync/sync-web/src/main/java/com/liferay/sync/web/internal/constants/SyncWebKeys.java
@@ -12,22 +12,13 @@
  * details.
  */
 
-package com.liferay.sync.upgrade.v1_0_3;
-
-import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.util.StringUtil;
+package com.liferay.sync.web.internal.constants;
 
 /**
  * @author Shinn Lok
  */
-public class UpgradeSchema extends UpgradeProcess {
+public interface SyncWebKeys {
 
-	@Override
-	protected void doUpgrade() throws Exception {
-		String template = StringUtil.read(
-			UpgradeSchema.class.getResourceAsStream("dependencies/update.sql"));
-
-		runSQLTemplateString(template, false, false);
-	}
+	public static final String IP_GEOCODER = "IP_GEOCODER";
 
 }
