+++ modules/apps/sync/sync-web/src/main/java/com/liferay/sync/web/internal/portlet/SyncAdminPortlet.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sync.web.portlet;
+package com.liferay.sync.web.internal.portlet;
 
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
