+++ modules/apps/sync/sync-web/src/main/resources/META-INF/resources/init.jsp
@@ -59,7 +59,7 @@ page import="com.liferay.sync.oauth.helper.SyncOAuthHelperUtil" %><%@
 page import="com.liferay.sync.service.SyncDeviceLocalServiceUtil" %><%@
 page import="com.liferay.sync.service.configuration.SyncServiceConfigurationKeys" %><%@
 page import="com.liferay.sync.service.configuration.SyncServiceConfigurationValues" %><%@
-page import="com.liferay.sync.web.constants.SyncWebKeys" %>
+page import="com.liferay.sync.web.internal.constants.SyncWebKeys" %>
 
 <%@ page import="java.util.ArrayList" %><%@
 page import="java.util.LinkedHashMap" %><%@
